+++ src/test/java/BaseTest.java
@@ -3,6 +3,7 @@ import com.codeborne.selenide.Selenide;
 import io.github.bonigarcia.wdm.WebDriverManager;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
+import pages.musicPage.pageElements.UpperMusicToolBar;
 
 public class BaseTest {
